refactor: move noscript to common
